\name{color}
\alias{color}
\docType{data}
\title{
  Two Cross-Classifications of Eye Color by Hair Color
}
\description{
  Three-way table crossing eye color and hair color in two places, Caithness and Aberdeen.
  This table is used by Becker and Clogg (1989) to illustrate several log-multiplicative
  row-column association models, with and without layer effect.
}
\usage{
data(color)
}
\references{
  Becker, M.P., and Clogg, C.C. (1989). Analysis of Sets of Two-Way Contingency Tables Using
    Association Models. \emph{J. of the Am. Stat. Association} 84(405), 142-151.
}
\examples{
  ## see \link{\code{rc}}
}
\keyword{datasets}
