\name{criminal}
\alias{criminal}
\docType{data}
\title{
  Dropped Criminal Charges, Denmark, 1955-1958
}
\description{
  Number of men aged 15-19 charged with a criminal case for whom charges
  were dropped: Denmark, 1955-1958. This two-way table is used by
  Goodman (1991) to illustrate a log-multiplicative row-column model
  with one dimension. It was used before by Rasch (1966), Christiansen
  and Stene (1969), and Andersen (1986, 1990).
}
\usage{data(criminal)}
\references{
  Goodman, L.A. (1991). Measures, Models, and Graphical Displays in the Analysis of
    Cross-Classified Data. \emph{J. of the Am. Stat. Association} 86(416), 1086,
    Table 1.
}
\examples{
  ## see ?rc
}
\keyword{datasets}
