\name{gss}
\alias{gss}
\docType{data}
\title{
  Education and Occupational Attainment Among Women
}
\description{
  Two-way table crossing education and occupational attainment among women
  from the General Social Survey: United States, 1985-1990. This table is used
  by Wong (2010) to illustrate a log-multiplicative row-column model with two dimensions.
}
\usage{data(gss)}
\references{
  Wong, R.S-K. (2010). Association models. SAGE. 32, Table 2.3 B.
}
\examples{
  ## see \link{\code{rc}} and \link{\code{plot.rc}}
}
\keyword{datasets}
