\name{gss7590}
\alias{gss7590}
\docType{data}
\title{
  Education and Occupational Attainment Among White Men and Women in the United States, 1975-1990
}
\description{
  Three-way table crossing education and occupational attainment by sex and period among white
  men and women from the General Social Survey: United States, 1975-1980 and 1985-1990. This
  table is used by Wong (2010) to illustrate log-multiplicative row-column models with three dimensions.
}
\usage{data(gss7590)}
\references{
  Wong, R.S-K. (2010). Association models. SAGE. 32, Table 4.3.
}
\examples{
  ## see ?rcL and ?plot.rcL
}
\keyword{datasets}
