\name{gss88}
\alias{gss88}
\docType{data}
\title{
  Major Occupation by Years of Schooling in the United States, 1988
}
\description{
  Two-way table crossing occupational attainment and years of education among persons aged
  20 through 64 from the General Social Survey: United States, 1988. This table is used
  by Clogg and Shihadeh (1994) to illustrate a log-multiplicative row-column model with one
  dimension.
}
\usage{data(gss88)}
\references{
  Clogg, C.C., and Shihadeh, E.S. (1994). Statistical Models for Ordinal Variables. Sage: Advanced
      Quantitative Techniques in the Social Sciences (4), Table 3.1, p. 40.
}
\examples{
  ## see ?rc
}
\keyword{datasets}
