\name{hg16}
\alias{hg16}
\docType{data}
\title{
  Son's occupation by father's occupation for 16 countries in the 1960s and 1970s
}
\description{
  Three-way mobility table assembled by Hazelrigg and Garnier (1976), and used by Zhou (2015)
  to illustrate the shrinkage estimation of the log-odds ratios and of the Altham index.
}
\usage{data(hg16)}
\references{
  Hazelrigg, L.E., Garnier, M.A (1976). Occupational Mobility in Industrial Societies:
      A Comparative Analysis of Differential Access to Occupational Ranks in Seventeen Countries.
      \emph{American Sociological Review} 41: 498-511.

  Zhou, X. (2015). Shrinkage Estimation of Log-Odds Ratios for Comparing Mobility Tables.
      \emph{Sociological Methodology} 45(1):33-63.
}
\examples{
  ## see ?iac
}
\keyword{datasets}
