\name{ocg1973}
\alias{ocg1973}
\docType{data}
\title{Intergenerational Mobility in the United States, 1973}
\description{
  Mobility table for the United States from the 1973 Occupational Changes in a Generation (OCG-II)
  survey. This table has been used by Yamaguchi (1987, 1990), Xie (1992) and van der Heijden &
  Mooijaart (1995).
}
\usage{data(ocg1973)}
\references{
  Yamaguchi, K. (1990). Some Models for the Analysis of Asymmetric Association in Square
  Contingency Tables with Ordered Categories. \emph{Sociological Methodology} 20, 181-212.

  van der Heijden, P.G.M., and Mooijaart, A. (1995). Some new log bilinear models for the analysis
    of asymmetry in a square contingency table. \emph{Sociol. Methods and Research} 24, 7-29.}
\examples{
  ## see \link{\code{yskew}}, \link{\code{hmskew}} and \link{\code{plot.hmskew}}
}
\keyword{datasets}
